pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "epfl-ldap>=0.0.6",
    "faiss-cpu>=1.10.0",
    "fiona>=1.10.1",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "geopandas>=1.0.1",
    "gunicorn>=23.0.0",
    "langchain>=0.3.23",
    "openai>=1.72.0",
    "psycopg2-binary>=2.9.10",
    "python-ldap>=3.4.4",
    "shapely>=2.1.0",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "numpy>=2.2.4",
    "langchain-community>=0.3.21",
    "tiktoken>=0.9.0",
]
